feat (wsconfig): allowing variable replacements in the wsconfig type endpoints
Generally in different environments our web service endpoint are different.  To facilitate this, I've added variable replacements to wsconfig type endpoint.
The syntax is to include  into the endpoint urls.  This is parsed out when the webservice call is made and the replacement is done
using variable_get(variable_name, 'default value');