feat (remove REST): remove the REST connector
We're removing the stub and the TODO for the REST connector as it is unlikely to be implemented.
This is because it is likely no long will use it, as REST services aren't really a defined protocol, so you have setup something custom anyway, so we're going to focus on making that easy to do.